+++ .golangci.yml
@@ -1,8 +1,11 @@
 run:
   modules-download-mode: vendor
-
+  skip-dirs:
+    - vendor
+    - e2e
 # Run only staticcheck for now. Additional linters will be enabled one-by-one.
 linters:
   enable:
   - staticcheck
+  - goimports
   disable-all: true
