+++ pkg/graphitesample/graphitesample.go
@@ -1,0 +1,36 @@
+// Copyright 2020 The Prometheus Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package graphitesample
+
+import (
+	"fmt"
+	"time"
+
+	"github.com/prometheus/client_golang/prometheus"
+)
+
+// GraphiteSample represents a graphite metric sample
+type GraphiteSample struct {
+	OriginalName string
+	Name         string
+	Labels       prometheus.Labels
+	Help         string
+	Value        float64
+	Type         prometheus.ValueType
+	Timestamp    time.Time
+}
+
+func (s GraphiteSample) String() string {
+	return fmt.Sprintf("%#v", s)
+}
