+++ pkg/metricmapper/mapper.go
@@ -1,0 +1,26 @@
+// Copyright 2020 The Prometheus Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package metricmapper
+
+import (
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/statsd_exporter/pkg/mapper"
+)
+
+// MetricMapper is an interface for mapper methods
+type MetricMapper interface {
+	GetMapping(string, mapper.MetricType) (*mapper.MetricMapping, prometheus.Labels, bool)
+	InitFromFile(string, int, ...mapper.CacheOption) error
+	InitCache(int, ...mapper.CacheOption)
+}
